buildTemplate moved to utils
