+++ src/core.ts
@@ -1,3 +1,4 @@
+import path from 'path';
 import { getTemplateBase } from './utils/getTemplateBase';
 import { getTranslationKeys } from './utils/getTranslationKeys';
 import { getPagesKeys } from './utils/getPagesKeys';
@@ -5,8 +6,7 @@ import { loadTranslation } from './utils/loadTranslation';
 import { composeData } from './utils/composeData';
 import { getFileType } from './utils/getFileType';
 import { getTemplateFile } from './utils/getTemplateFile';
-import path from 'path';
-import buildTemplate from '.';
+import buildTemplate from './utils/buildTemplate';
 import { minifyBuildContent } from './utils/minifyBuildContent';
 import { ENV_DEVELOPMENT } from './app/defines';
 import type { Environment, ConfigFile } from './types';
@@ -41,9 +41,6 @@ export const core = ({
     formatter,
     minify,
 }: CoreOptions): CoreResult => {
-
-    console.log("core: options", { config, env, lang, page, cwd, formatter, minify });
-
     const files: CoreFile[] = [];
 
     const templateBase = getTemplateBase(cwd, config.options.src);
@@ -109,8 +106,6 @@ export const core = ({
             files.push(file);
         });
     });
-
-    console.log("core: files", files);
 
     return files;
 };
